feat(mempool_watcher): add configuration flags to enable/disable fetchers
This commit introduces two new configuration flags, `EthPendingTxsEnabled` and `TxPoolContentEnabled`, to the `MempoolWatcher` configuration. These flags control whether the periodic fetching of `eth_pendingTransactions` and `txpool_content` is enabled, respectively.

Previously, both fetchers were always started when the `MempoolWatcher` was started. This change allows for more granular control over which methods are used to fetch pending transactions, enabling users to disable specific fetchers if they are not needed or causing issues.

A new test case `TestFetchersBasedOnConfig` has been added to verify that the fetcher goroutines are started or not based on the values of these configuration flags.